Use qlonglong for file size in C++
This does not yet fix the bug:
The binary protocol of the ioslaves will have to be changed.
I have avoided this so far, but plan to do it.

As well, this change needs to be evaluated for MySQL as discussed by Francesco.

CCBUG: 276920

+++ libs/database/collectionscanner.cpp
@@ -883,7 +883,7 @@ void CollectionScanner::scanFileNormal(const QFileInfo& fi, const ItemScanInfo& 
     {
         // if the file need not be scanned because of modification, update the hash
         if (modificationDateEquals(fi.lastModified(), scanInfo.modificationDate)
-            && (int)fi.size() == scanInfo.fileSize)
+            && fi.size() == scanInfo.fileSize)
         {
             QString oldHash = scanInfo.uniqueHash;
             QString newHash = scanFileUpdateHash(fi, scanInfo);
@@ -897,7 +897,7 @@ void CollectionScanner::scanFileNormal(const QFileInfo& fi, const ItemScanInfo& 
     }
 
     if (!modificationDateEquals(fi.lastModified(), scanInfo.modificationDate)
-        || (int)fi.size() != scanInfo.fileSize)
+        || fi.size() != scanInfo.fileSize)
     {
         scanModifiedFile(fi, scanInfo);
     }
