+++ libs/database/downloadhistory.cpp
@@ -32,7 +32,7 @@ namespace Digikam
 {
 
 DownloadHistory::Status DownloadHistory::status(const QString& identifier, const QString& name,
-        int fileSize, const QDateTime& date)
+        qlonglong fileSize, const QDateTime& date)
 {
     int id = DatabaseAccess().db()->findInDownloadHistory(identifier, name, fileSize, date);
 
@@ -47,7 +47,7 @@ DownloadHistory::Status DownloadHistory::status(const QString& identifier, const
 }
 
 void DownloadHistory::setDownloaded(const QString& identifier, const QString& name,
-                                    int fileSize, const QDateTime& date)
+                                    qlonglong fileSize, const QDateTime& date)
 {
     DatabaseAccess().db()->addToDownloadHistory(identifier, name, fileSize, date);
 }
